feat: migrate database from SQLite to PostgreSQL
BREAKING CHANGE: Database engine changed from SQLite/libsql to PostgreSQL.

## Summary
Complete migration of the GEA EOR SaaS platform from SQLite to PostgreSQL
for production-grade deployment with proper concurrency support.

## Changes

### Schema (drizzle/schema.ts, drizzle/aor-schema.ts)
- Converted all 63 tables from sqliteTable to pgTable
- integer({ mode: 'boolean' }) → boolean()
- integer({ mode: 'timestamp_ms' }) → timestamp({ withTimezone: true })
- integer().primaryKey({ autoIncrement: true }) → serial().primaryKey()
- text({ mode: 'json' }) → jsonb()
- text({ length: N }) → varchar({ length: N })
- real() → doublePrecision()

### Database Connection (server/services/db/connection.ts, server/db.ts)
- Replaced @libsql/client with postgres.js driver
- Replaced drizzle-orm/libsql with drizzle-orm/postgres-js
- Added connection pooling configuration

### Raw SQL Fixes
- CAST(x AS REAL) → CAST(x AS numeric) (17 occurrences)
- strftime('%Y-%m', col) → to_char(col::date, 'YYYY-MM') (4 occurrences)
- substr(col, 1, N) → substring(col from 1 for N) (2 occurrences)
- db.run() → db.execute() (3 occurrences)

### Auto-Migration (server/autoMigrate.ts)
- Rewrote from SQLite PRAGMA-based to PostgreSQL information_schema
- ALTER COLUMN DROP NOT NULL replaces table recreation workaround

### Docker Deployment
- Added PostgreSQL 16 service to docker-compose.prod.yml
- Updated docker-entrypoint.sh to use postgres.js for migrations
- Updated Dockerfile.prod comments

### Scripts & Tests
- Updated all seed/migrate/verify scripts to use postgres.js
- Updated all test files to use PostgreSQL connection strings
- Added data migration script: scripts/migrate-sqlite-to-pg.ts

### Configuration
- drizzle.config.ts: dialect changed to 'postgresql'
- .env.example: Updated DATABASE_URL format and examples
- server/_core/env.ts: Relaxed DATABASE_URL validation

+++ server/routers/adminDashboard.ts
@@ -888,7 +888,7 @@ export const adminDashboardRouter = router({
       })
       .from(customerWallets)
       .leftJoin(customers, eq(customerWallets.customerId, customers.id))
-      .where(sql`CAST(${customerWallets.balance} AS REAL) < 0`);
+      .where(sql`CAST(${customerWallets.balance} AS numeric) < 0`);
 
     // ── 9. Expired Onboarding Invites (still pending) ──
     const expiredInvites = await db
